Attempt to optimize GUI
The main program cannot communicate with GUI.
If the GUI read output file while the main program write it, user will see white screen on GUI, because the output file is an imcomplete HTML file.
Attempt is to build two iframe stack together. If one is empty, the user can see the other one. But this solution gives bad result in experiment.